Incorporated code from CanCPL_CICE to allow NEMO to be coupled with the CCCma AGCM.
This will form the starting point for using LIM2 as the sea ice model in NEMO for a coupled
configuration with NEMO coupled to the CCCma AGCM via a stand alone coupler.
Certain CICE related code was not carried over into this commit, namely the CICE source
files from nemo/NEMO/CICE_SRC_5 and changes made to nemo/NEMO/OPA_SRC/SBC/sbcice_cice.F90.
These files will need to be added if (when) CICE becomes functional in this coupled env.

The commit in ~acrnsrc/ocean_dir/nemo-CCCma.git at which this code was cloned is
8ff72767ed7124e2ea915c079d3a26a77a50aede. This was the head of the develop branch
at the time. It is this revision to which the coupler related changes were added.

The commit from the nemo_cancpl_dev_CICE repo used for coupler related changes is
6f7ba67a54722af5d4dcc4d956bc935bcab623ee.