Modifies p4zsed depth loops to correct bug
- Modifies euphotics zone loop for calcite removal to run from

   jk=1,jk_eud_cmoc-1

as opposed to

  jk=1,jk_eud_cmoc

In the latter case the index is specifying the modification of t-grid
variables below the euphotic zone (jk_eud_cmoc), which is defined on
the w-grid. I.e. this correct a logical error which did not properly
account for how the w and t grids were staggered.