Modified nemo_poslude to fix problems dumping/deleting files when operating in gcmjcl mode.
Need to parse the nemo restart file name to get the current month differently in gcmjcl mode
vs xnemo mode since this file name is different in the 2 cases.

Do not access certain files (e.g. rtd executables and grid description file) if they already
exist in the execution dir. This happens on all months after the first when running multiple
months in a single submission.