Add an ARCH configuration enabling debugging for XC40 Intel and Cray
In order to allow an executable to be debugged, some compiler settings need
to be set. Namely, optimisations should be turned off (otherwise loops may
not be visible) and the debugging symbols should be output. The Intel
version of this has been tested and works with Allinea Forge visual
debugger.