Modify build-nemo to allow building debugging executable
For debugging, the source files which are preprocessed in compile-nemo
need to be retained and stored in a directory visible at runtime (either on
hare or brooks). Additionally a modified ARCH file needs to be used to
ensure that the compiler options are suitable for debugging. The changes
to the build-nemo script enable this behaviour using the '-d' option.