Print control statements as checksums
NEMO by default can print out debugging statements when
ln_ctl=.true. These are printed as SUM of that array on every PE.
This is not as useful to debug the deviation of the model from its
expected state. A new runtime flag, ln_ctl_chksum, toggles whether
the original array sum is calculated or a chksum of the requested
array is output.