Calculate 'abiotic' alkalinity using OMIP protocol
The Orr et al. 2016 paper specifies that all abiotic tracer should
use an alkalinity defined as the 2297 umol/kg * S/\bar{S}, where
\bar{S} is the globally averaged surface salinity (if from a cold start)
or the globally and annually averaged surface salinity from the
previous year. These commits add in all the routines necessary to do this
calculation. The next step will be to remove the 'abiotic' alkalinity
tracer to instead use this formulation