Fix array used for DI14C surface flux
DI14C somehow was using the tra array instead of trn in the calculation
of zh2co3r. Additionally, the model would could throw a floating point
error due to a potential divide by zero. This has been replaced with
a MAX statement in the denominator and a multiplication by tmask