Separate interpolation of 14C and CO2 boundary condition
The boundary value for CO2 and 14C were originally interpolated in
time or set to a constant depending on the value of ln_co2int. For
coupled configurations the value of ln_co2int does not change the
behavior of the model because the value seen during gas exchange
will have been overwritten by the pCO2 value from the coupler.

To avoid ambiguity in behavior for C14, a new flag ln_c14int has
been added so that the time interpolation can be separately controlled
for CO2 and 14C.