Fix problems with time interpolation in ocean carbon fields
The logical test for finding whether or not to interpolate was wrong
typically meaning that CO2 and C14 were reading the very first time
level of the file. Additionally the order of dimenisons for the CO2
was flipped when reading in the fields (C vs Fortran convention).
The time interpolation should have only affected the carbon fields
because the CFCs and SF6 use their own function.