Add option to distribute ice salt flux vertically
To try to prevent salinity from going to unrealistically high values in
regions of sea-ice formation and strong horizontal divergence, a new
option ln_vertsflx has been added. This distributes all salt fluxes from
the ice in the vertical by depositing a portion of the flux into each
layer weighted by the thickness of the layer over the depth of the
diagnosed boundary layer. This could approximately be interpreted as a
crude approximation of mixed layer convection due to brine rejection.