Revise salt plume parameterizaton
The original implementation of the salt plume parameterization after
Nguyen et al. 2009, naively implemented the coefficient stated in the
paper. However, the normalization factor is not appropriate for a
discrete layer formulation. In addition, the salt plume parameterization
should only be applied when ice is being formed. Otherwise, if the
ice is melting, it should only enter at the surface (maybe at the
keel depth of the ice). Lastly, this adds diagnostics of the 3d
tendency associated with the salt plume and its thickness