Add option to fix the sign of salt flux
The salt flux due to sea-ice formation/melt uses the sign convention
that the flux is negative during formation. However, when this salt flux
is applied to the ocean salinity, the sign is not switched. Hence,
sea-ice growth ends up freshening the ocean. This commit adds a new
runtime parameter nn_sfx_sign and sets the default value to be -1 to be
consistent with the ice model's convention for sfx