Bugfix: Calculation of coupling time
The subroutine query_start_cpl2ocn is used to determine when the coupler
is ready to send fields to the ocean. Due to some changes in the NEMO4
code the length of time elapsed was not being calculated correctly and
so the model was hanging. This commit fixes this bug and allows the
model to continue integrating. It however still is crashing in the
ocean, but it is not clear whether this is because the fields are
incorrect, or if this is related to the ocean-only instabilities.