Fix behavior of NEMO with ln_vertspp off
There was an inconsistency in the way that the mass flux from ice
melt/formation was being handled in the case where ln_verstpp was
turned off. Essentially, this flux was being added to emp (as was
originally done in NEMO), but then also being applied again in the
dynamics solver and the tracer advection portion of the code;
essentially double counting the flux. @rdy001 noticed that his caused
both MOC and MHT to decline significantly.

The changes in this commit restores the original behavior of the code as
determined by the switch ln_vertspp.