Fix reconstruct and scatter routines
Fixes the reconstruct routine and updates the mppscatter routine for use
with non-uniform decompositions of the subdomains. This allows the
coupled model to run for 10 days. Note there still remains a bug in
NEMO4 where some subdomains are incorrectly flagged as land points.
However, this seems to come from the original NEMO codebase.