Avoid divide by zero errors in xios client
In the coupled mode, when the client is embedded within NEMO and is not
a standalone process, the timing output was throwing a divide by zero
error. Additionally, there seems to be a bug in the Intel compiler which
leads to optimizations inappropriately skipping a divide by zero check
in the averaging routine in the presence of missing values. This adds a
compiler directive to avoid optimizing the loop.