Fixes a bug in the closea generation
The coupled NEMO4 configuration was still crashing when the new
closea_mask was being used. This was tracked down to two problems:
1. The masking in closea.F90 incorrectly does not seem to be correct
   This might be because the tmask has yet to be initialized or
   something odd going on in the halos
2. The fill values for a masked variable is used when a variable
   is read leading to unrealistically large values. This was
   fixed by modifying the fill value in make_closea_masks.py

Additionally, this commit reverts a previous change that forced the
generation of the empr and runoff masks. This was not necessary and
indeed can lead to additional problems.